# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "wday2",
        "description": "Backend sistema de gestión de personal",
        "contact": {
            "email": "totomarion@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Crear de usuario",
                "operationId": "signupAuthController",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthSignupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Login de usuario",
                "operationId": "loginAuthController",
                "parameters": [
                    {
                        "name": "login",
                        "in": "query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthLoginRequest"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario logueado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Logout de usuario",
                "operationId": "logoutAuthController",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/user": {
            "get": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Datos del usuario",
                "operationId": "user",
                "responses": {
                    "200": {
                        "description": "Usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/password/create": {
            "post": {
                "tags": [
                    "PasswordResetController"
                ],
                "summary": "Restablecer contraseña del usuario",
                "operationId": "createPasswordResetController",
                "parameters": [
                    {
                        "name": "login",
                        "in": "query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PasswordResetCreateRequest"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contraseña restablecida",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/password/find": {
            "get": {
                "tags": [
                    "PasswordResetController"
                ],
                "summary": "Si el token está activo, la respuesta contendrá los datos de éste",
                "operationId": "logoutAuthController",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sesión cerrada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "PasswordResetController"
                ],
                "summary": "Si token activo devuelve usuario",
                "operationId": "resetPasswordResetController",
                "parameters": [
                    {
                        "name": "login",
                        "in": "query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PasswordResetResetRequest"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "AuthLoginRequest": {
                "title": "AuthLoginRequest",
                "description": "Login Request",
                "properties": {
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "format": "password"
                    },
                    "remember_me": {
                        "description": "Recordar conexión del usuario",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "AuthSignupRequest": {
                "title": "AuthSignupRequest",
                "description": "Signup Request",
                "properties": {
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "format": "password"
                    },
                    "password_confirmation": {
                        "description": "Confirmar contraseña del usuario",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
            "PasswordResetCreateRequest": {
                "title": "PasswordResetCreateRequest",
                "description": "Password Reset Create Request",
                "properties": {
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            },
            "PasswordResetResetRequest": {
                "title": "PasswordResetResetRequest",
                "description": "Password Reset Request",
                "properties": {
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "format": "password"
                    },
                    "token": {
                        "description": "Token del usuario",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Representación del usuario",
                "properties": {
                    "id": {
                        "description": "Id del usuario",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del usuario",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "description": "Cuando el usuario verifica su email",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "active": {
                        "description": "Si usuario se encuentra activo",
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}